Extracted scope operations into a concern
